ElevateU-1/backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-environ>=0.11
django-cors-headers>=4.4
psycopg2-binary>=2.9
gunicorn>=22.0
djangorestframework-simplejwt>=5.2
Pillow>=9.0
boto3>=1.20
celery>=5.2
redis>=4.0